refactor: 파일, 클래스 명 수정 및 문자열을 상수로 수정, ManagerImpl은 리스너 2개로 변경
현재 버그가 회원여부는 true가 전달이 되는데 닉네임 화면 즉 온보딩 과정을 거치는 버그가 있습니다. 그러고 앱을 종료 한 다음에 다시 실행하면 홈이 켜집니다.

리스너를 다시 2개로 수정하면서 제대로 된 화면 전환이 되도록 했습니다.
