+++ common/src/main/java/com/paykidscompose/common/model/AuthStatusManager.kt
@@ -2,7 +2,7 @@ package com.paykidscompose.common.model
 
 import kotlinx.coroutines.flow.Flow
 
-abstract class TokenManager {
+abstract class AuthStatusManager {
     abstract fun getIsRegistered() : Flow<Boolean>
     abstract fun getToken() : Flow<String>
 }
