+++ common/src/main/java/com/paykidscompose/common/usecase/authentication/CheckUserCompletionStatusUseCase.kt
@@ -1,16 +1,16 @@
 package com.paykidscompose.common.usecase.authentication
 
 import com.paykidscompose.common.enums.EntryPoint
-import com.paykidscompose.common.model.TokenManager
+import com.paykidscompose.common.model.AuthStatusManager
 import com.paykidscompose.common.usecase.base.FlowUseCase
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 
 class CheckUserCompletionStatusUseCase(
-    private val tokenManager: TokenManager
+    private val authStatusManager: AuthStatusManager
 ) : FlowUseCase<Unit, EntryPoint>() {
     override fun execute(params: Unit?): Flow<EntryPoint> {
-        return tokenManager.getToken().combine(tokenManager.getIsRegistered()) { token, isRegistered ->
+        return authStatusManager.getToken().combine(authStatusManager.getIsRegistered()) { token, isRegistered ->
             if (token.isBlank()) {
                 EntryPoint.LOGIN
             } else if (isRegistered) {
