+++ data/src/main/java/com/paykidscompose/data/util/Constants.kt
@@ -3,3 +3,4 @@ package com.paykidscompose.data.util
 const val ACCESS_TOKEN = "accessToken"
 const val REFRESH_TOKEN = "refreshToken"
 const val INTERCEPTOR_HEADER = "Authorization"
+const val USER_REGISTERED = "userRegistered"
